Move components into their own folders
